refactor(poller): keep Infrastructure/Dbal for what depends on the connector
The mapping and its exception sat under Infrastructure/Dbal, first in an
Exception/ subfolder, then prefixed Dbal to match their 21 neighbours. Neither
placement held up: the prefix in that folder marks a dependency on the connector,
and these two have none. They inject no connection, import neither Doctrine nor
Adaptation\Database, and handle nothing but strings and a domain enum — swapping
Doctrine for PDO would not change a line of them. What they know is the shape of
the data: that the column encodes ZMQ as '1' and PullWSS as '4'.

So they move up to Infrastructure/, unprefixed, next to InstallationCommandFactory
and PollerInstallationCommandFactory which are there for the same reason. Dbal/ is
left holding only classes that touch the connector, all of them prefixed.

The rule this applies — prefix if and only if the class depends on the connector,
otherwise one level up — leaves the nine Dbal*Transformer of this module and the
two in Security in violation: they implement the generic TransformerInterface and
receive an already-read row, never a connection. Renaming eleven files across two
modules would bury a bug fix under a convention refactor, so they are tracked
separately by MON-207786.

Two mechanical consequences of the move: the test dropped a directory, so its
dirname(__DIR__, 6) became 5, and the two Dbal classes that use the mapping now
import it explicitly instead of resolving it in their own namespace.

Refs: MON-207613
